requirements.txt
asgiref~=3.6.0
chardet~=4.0.0
cssselect2==0.7.0
Django~=4.2.1
django-filer==3.1.3
django-filter==24.2
django-polymorphic==3.1.0
easy-thumbnails==2.8.5
lxml~=4.6.3
pillow~=8.2.0
reportlab==4.2.0
sqlparse~=0.4.4
svglib==1.5.1
tinycss2==1.3.0
tzdata==2024.1
webencodings==0.5.1

cffi~=1.14.5
defusedxml~=0.7.1
future~=0.18.2
pip~=21.1.1
attrs~=21.2.0
cryptography~=40.0.2
docutils~=0.17.1
filelock~=3.0.12
beautifulsoup4~=4.9.3
PyYAML~=5.4.1
numpy~=1.20.1
requests~=2.25.1
setuptools~=56.2.0
django-admin-sortable2~=2.2.1